fix: start separating display from util, migrating execution
Starting with a components folder for a crew, and contemplating moving their respective agent and task definitions into a single file. Having devs submit a single file is easier than this change here, that change there.

Also removing the class names from the agents file broke the execution here, as they were imported that way (and it looks much cleaner). Less of a problem if defined in same file though.